+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed an issue on Android where the notification background would be displayed in low resolution on recent Android versions.
 
+### Added
+
+- Added support on iOS for 'album' and 'artist' in the nowPlayingInfo.
+
 ## [7.8.0] - 24-08-09
 
 ### Added