+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [7.8.1] - 24-08-14
+
+### Fixed
+
+- Fixed an issue on Android where the notification background would be displayed in low resolution on recent Android versions.
+
+### Added
+
+- Added support on iOS for 'album' and 'artist' in the nowPlayingInfo.
+
 ## [7.8.0] - 24-08-09
 
 ### Added